[gtk] adding methods and signals for usermedia communication between webkit and browser
When the user make a getUserMedia call, webview will emit a sigial (user-media-requested)
to the application containing then request itself (webkitwebusermediarequest) and two
list with the available audio and video devices (webkitwebusermedialist).

The application must show the list to the user and ask permission to use one of
those devices (or reject the request).

In case the user accepts the request, the application must mark the device as
selected (using webkit_web_user_media_list_select_item) and then call
webkit_web_view_accept_user_media_request which will handle the request properly